@@ -38,11 +71,10 @@
   },
   "author": {
     "name": "MinIO, Inc.",
-    "email": "",
     "url": "https://min.io"
   },
   "engines": {
-    "node": ">8  <=19"
+    "node": "^16 || ^18 || >=20"
   },
   "license": "Apache-2.0",
   "bugs": {
@@ -51,53 +83,60 @@
   },
   "homepage": "https://github.com/minio/minio-js#readme",
   "dependencies": {
-    "async": "^3.1.0",
-    "block-stream2": "^2.0.0",
-    "browser-or-node": "^1.3.0",
+    "async": "^3.2.4",
+    "block-stream2": "^2.1.0",
+    "browser-or-node": "^2.1.1",
     "buffer-crc32": "^0.2.13",
-    "crypto-browserify": "^3.12.0",
-    "es6-error": "^4.1.1",
-    "fast-xml-parser": "^4.1.3",
+    "fast-xml-parser": "^4.2.2",
     "ipaddr.js": "^2.0.1",
     "json-stream": "^1.0.0",
     "lodash": "^4.17.21",
-    "mime-types": "^2.1.14",
-    "mkdirp": "^0.5.1",
-    "query-string": "^7.1.1",
-    "through2": "^3.0.1",
+    "mime-types": "^2.1.35",
+    "mkdirp": "^0.5.6",
+    "query-string": "^7.1.3",
+    "through2": "^4.0.2",
     "web-encoding": "^1.1.5",
-    "xml": "^1.0.0",
+    "xml": "^1.0.1",
     "xml2js": "^0.5.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.12.10",
-    "@babel/preset-env": "^7.12.10",
-    "@types/lodash": "^4.14.192",
-    "@types/node": "^18.15.11",
-    "@typescript-eslint/eslint-plugin": "^5.57.1",
-    "@typescript-eslint/parser": "^5.57.1",
-    "babelify": "^10.0.0",
-    "browserify": "^16.5.2",
-    "chai": "^4.2.0",
-    "eslint": "^8.37.0",
+    "@babel/core": "^7.21.8",
+    "@babel/plugin-transform-modules-commonjs": "^7.21.5",
+    "@babel/preset-env": "^7.21.5",
+    "@babel/preset-typescript": "^7.21.5",
+    "@babel/register": "^7.21.0",
+    "@nodelib/fs.walk": "^1.2.8",
+    "@types/async": "^3.2.20",
+    "@types/lodash": "^4.14.194",
+    "@types/mime-types": "^2.1.1",
+    "@types/node": "^20.1.0",
+    "@types/xml": "^1.0.8",
+    "@types/xml2js": "^0.4.11",
+    "@typescript-eslint/eslint-plugin": "^5.59.2",
+    "@typescript-eslint/parser": "^5.59.2",
+    "@upleveled/babel-plugin-remove-node-prefix": "^1.0.5",
+    "babel-plugin-replace-import-extension": "^1.1.3",
+    "babel-plugin-transform-replace-expressions": "^0.2.0",
+    "chai": "^4.3.7",
+    "dotenv": "^16.0.3",
+    "eslint": "^8.40.0",
     "eslint-config-prettier": "^8.8.0",
+    "eslint-import-resolver-typescript": "^3.5.5",
+    "eslint-plugin-import": "^2.27.5",
     "eslint-plugin-simple-import-sort": "^10.0.0",
-    "gulp": "^4.0.2",
-    "gulp-babel": "^8.0.0",
-    "gulp-eslint": "^6.0.0",
-    "gulp-if": "^3.0.0",
-    "gulp-mocha": "^8.0.0",
-    "gulp-sourcemaps": "^2.6.5",
+    "eslint-plugin-unicorn": "^47.0.0",
+    "eslint-plugin-unused-imports": "^2.0.0",
     "husky": "^8.0.3",
-    "lint-staged": "^13.2.1",
-    "mocha": "^9.2.0",
-    "mocha-steps": "^1.1.0",
-    "nock": "^13.2.2",
-    "prettier": "^2.8.7",
-    "source-map-support": "^0.5.13",
-    "split-file": "^2.2.2",
-    "superagent": "^5.1.0",
-    "uuid": "^3.4.0"
+    "lint-staged": "^13.2.2",
+    "mocha": "^10.2.0",
+    "mocha-steps": "^1.3.0",
+    "nock": "^13.3.1",
+    "prettier": "^2.8.8",
+    "source-map-support": "^0.5.21",
+    "split-file": "^2.3.0",
+    "superagent": "^8.0.1",
+    "typescript": "^5.0.4",
+    "uuid": "^9.0.0"
   },
   "keywords": [
     "api",
