@@ -14,35 +14,34 @@
  * limitations under the License.
  */
 
+import * as fs from 'node:fs'
+import * as Http from 'node:http'
+import * as Https from 'node:https'
+import * as path from 'node:path'
+import * as Stream from 'node:stream'
+
 import async from 'async'
 import BlockStream2 from 'block-stream2'
-import fs from 'fs'
-import Http from 'http'
-import Https from 'https'
 import _ from 'lodash'
 import mkdirp from 'mkdirp'
-import path from 'path'
-import querystring from 'query-string'
-import Stream from 'stream'
+import * as querystring from 'query-string'
 import { TextEncoder } from 'web-encoding'
 import Xml from 'xml'
 import xml2js from 'xml2js'
 
-import CredentialProvider from './CredentialProvider'
-import * as errors from './errors.js'
-import extensions from './extensions'
+import { CredentialProvider } from './CredentialProvider.ts'
+import * as errors from './errors.ts'
+import { extensions } from './extensions.js'
+import { CopyDestinationOptions, CopySourceOptions, DEFAULT_REGION } from './helpers.ts'
+import { CopyConditions } from './internal/copy-conditions.ts'
 import {
   calculateEvenSplits,
-  CopyDestinationOptions,
-  CopySourceOptions,
-  DEFAULT_REGION,
   extractMetadata,
   getScope,
   getSourceVersionId,
   getVersionId,
   insertContentType,
   isAmazonEndpoint,
-  isArray,
   isBoolean,
   isFunction,
   isNumber,
@@ -56,30 +55,34 @@ import {
   isValidPort,
   isValidPrefix,
   isVirtualHostStyle,
-  LEGAL_HOLD_STATUS,
   makeDateLong,
   PART_CONSTRAINTS,
   partsRequired,
   pipesetup,
   prependXAMZMeta,
-  promisify,
   readableStream,
-  RETENTION_MODES,
-  RETENTION_VALIDITY_UNITS,
   sanitizeETag,
   toMd5,
   toSha256,
   uriEscape,
   uriResourceEscape,
-} from './helpers.js'
-import { NotificationConfig, NotificationPoller } from './notification'
-import ObjectUploader from './object-uploader'
-import { getS3Endpoint } from './s3-endpoints.js'
-import { postPresignSignatureV4, presignSignatureV4, signV4 } from './signing.js'
-import * as transformers from './transformers'
-import { parseSelectObjectContentResponse } from './xml-parsers'
+} from './internal/helper.ts'
+import { PostPolicy } from './internal/post-policy.ts'
+import { getS3Endpoint } from './internal/s3-endpoints.ts'
+import { LEGAL_HOLD_STATUS, RETENTION_MODES, RETENTION_VALIDITY_UNITS } from './internal/type.ts'
+import { NotificationConfig, NotificationPoller } from './notification.js'
+import { ObjectUploader } from './object-uploader.js'
+import { promisify } from './promisify.js'
+import { postPresignSignatureV4, presignSignatureV4, signV4 } from './signing.ts'
+import * as transformers from './transformers.js'
+import { parseSelectObjectContentResponse } from './xml-parsers.js'
 
-var Package = require('../../package.json')
+export * from './helpers.ts'
+export * from './notification.js'
+export { CopyConditions, PostPolicy }
+
+// will be replaced by bundler
+const Package = { version: process.env.MINIO_JS_PACKAGE_VERSION || 'development' }
 
 export class Client {
   constructor(params) {
@@ -564,7 +567,7 @@ export class Client {
         }
 
         this.checkAndRefreshCreds()
-        var authorization = signV4(reqOptions, this.accessKey, this.secretKey, region, date)
+        var authorization = signV4(reqOptions, this.accessKey, this.secretKey, region, date, sha256sum)
         reqOptions.headers.authorization = authorization
       }
       var req = this.transport.request(reqOptions, (response) => {
@@ -918,7 +921,7 @@ export class Client {
   // * `callback(err)` _function_: callback function is called with non `null` value in case of error
   removeIncompleteUpload(bucketName, objectName, cb) {
     if (!isValidBucketName(bucketName)) {
-      throw new errors.isValidBucketNameError('Invalid bucket name: ' + bucketName)
+      throw new errors.IsValidBucketNameError('Invalid bucket name: ' + bucketName)
     }
     if (!isValidObjectName(objectName)) {
       throw new errors.InvalidObjectNameError(`Invalid object name: ${objectName}`)
@@ -1908,7 +1911,7 @@ export class Client {
     if (!isValidBucketName(bucketName)) {
       throw new errors.InvalidBucketNameError('Invalid bucket name: ' + bucketName)
     }
-    if (!isArray(objectsList)) {
+    if (!Array.isArray(objectsList)) {
       throw new errors.InvalidArgumentError('objectsList should be a list')
     }
     if (!isFunction(cb)) {
@@ -2632,7 +2635,7 @@ export class Client {
     if (!isString(suffix)) {
       throw new TypeError('suffix must be of type string')
     }
-    if (!isArray(events)) {
+    if (!Array.isArray(events)) {
       throw new TypeError('events must be of type Array')
     }
     let listener = new NotificationPoller(this, bucketName, prefix, suffix, events)
@@ -2929,8 +2932,8 @@ export class Client {
     })
   }
 
-  /** Put lifecycle configuration on a bucket.
-  /** Apply lifecycle configuration on a bucket.
+  /**
+   * Apply lifecycle configuration on a bucket.
    * bucketName _string_
    * policyConfig _object_ a valid policy configuration object.
    * `cb(error)` _function_ - callback function with `err` as the error argument. `err` is null if the operation is successful.
@@ -3538,7 +3541,7 @@ export class Client {
     const me = this // many async flows. so store the ref.
     const sourceFilesLength = sourceObjList.length
 
-    if (!isArray(sourceObjList)) {
+    if (!Array.isArray(sourceObjList)) {
       throw new errors.InvalidArgumentError('sourceConfig should an array of CopySourceOptions ')
     }
     if (!(destObjConfig instanceof CopyDestinationOptions)) {
@@ -3849,137 +3852,3 @@ Client.prototype.setObjectLegalHold = promisify(Client.prototype.setObjectLegalH
 Client.prototype.getObjectLegalHold = promisify(Client.prototype.getObjectLegalHold)
 Client.prototype.composeObject = promisify(Client.prototype.composeObject)
 Client.prototype.selectObjectContent = promisify(Client.prototype.selectObjectContent)
-
-export class CopyConditions {
-  constructor() {
-    this.modified = ''
-    this.unmodified = ''
-    this.matchETag = ''
-    this.matchETagExcept = ''
-  }
-
-  setModified(date) {
-    if (!(date instanceof Date)) {
-      throw new TypeError('date must be of type Date')
-    }
-
-    this.modified = date.toUTCString()
-  }
-
-  setUnmodified(date) {
-    if (!(date instanceof Date)) {
-      throw new TypeError('date must be of type Date')
-    }
-
-    this.unmodified = date.toUTCString()
-  }
-
-  setMatchETag(etag) {
-    this.matchETag = etag
-  }
-
-  setMatchETagExcept(etag) {
-    this.matchETagExcept = etag
-  }
-}
-
-// Build PostPolicy object that can be signed by presignedPostPolicy
-export class PostPolicy {
-  constructor() {
-    this.policy = {
-      conditions: [],
-    }
-    this.formData = {}
-  }
-
-  // set expiration date
-  setExpires(date) {
-    if (!date) {
-      throw new errors.InvalidDateError('Invalid date : cannot be null')
-    }
-    this.policy.expiration = date.toISOString()
-  }
-
-  // set object name
-  setKey(objectName) {
-    if (!isValidObjectName(objectName)) {
-      throw new errors.InvalidObjectNameError(`Invalid object name : ${objectName}`)
-    }
-    this.policy.conditions.push(['eq', '$key', objectName])
-    this.formData.key = objectName
-  }
-
-  // set object name prefix, i.e policy allows any keys with this prefix
-  setKeyStartsWith(prefix) {
-    if (!isValidPrefix(prefix)) {
-      throw new errors.InvalidPrefixError(`Invalid prefix : ${prefix}`)
-    }
-    this.policy.conditions.push(['starts-with', '$key', prefix])
-    this.formData.key = prefix
-  }
-
-  // set bucket name
-  setBucket(bucketName) {
-    if (!isValidBucketName(bucketName)) {
-      throw new errors.InvalidBucketNameError(`Invalid bucket name : ${bucketName}`)
-    }
-    this.policy.conditions.push(['eq', '$bucket', bucketName])
-    this.formData.bucket = bucketName
-  }
-
-  // set Content-Type
-  setContentType(type) {
-    if (!type) {
-      throw new Error('content-type cannot be null')
-    }
-    this.policy.conditions.push(['eq', '$Content-Type', type])
-    this.formData['Content-Type'] = type
-  }
-
-  // set Content-Type prefix, i.e image/ allows any image
-  setContentTypeStartsWith(prefix) {
-    if (!prefix) {
-      throw new Error('content-type cannot be null')
-    }
-    this.policy.conditions.push(['starts-with', '$Content-Type', prefix])
-    this.formData['Content-Type'] = prefix
-  }
-
-  // set Content-Disposition
-  setContentDisposition(value) {
-    if (!value) {
-      throw new Error('content-disposition cannot be null')
-    }
-    this.policy.conditions.push(['eq', '$Content-Disposition', value])
-    this.formData['Content-Disposition'] = value
-  }
-
-  // set minimum/maximum length of what Content-Length can be.
-  setContentLengthRange(min, max) {
-    if (min > max) {
-      throw new Error('min cannot be more than max')
-    }
-    if (min < 0) {
-      throw new Error('min should be > 0')
-    }
-    if (max < 0) {
-      throw new Error('max should be > 0')
-    }
-    this.policy.conditions.push(['content-length-range', min, max])
-  }
-
-  // set user defined metadata
-  setUserMetaData(metaData) {
-    if (!isObject(metaData)) {
-      throw new TypeError('metadata should be of type "object"')
-    }
-    Object.entries(metaData).forEach(([key, value]) => {
-      const amzMetaDataKey = `x-amz-meta-${key}`
-      this.policy.conditions.push(['eq', `$${amzMetaDataKey}`, value])
-      this.formData[amzMetaDataKey] = value
-    })
-  }
-}
-
-export * from './helpers'
-export * from './notification'
