@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 
-import Crypto from 'crypto'
-import _ from 'lodash'
+import * as crypto from 'node:crypto'
 
-import * as errors from './errors.js'
-import { getScope, isArray, isNumber, isObject, isString, makeDateLong, makeDateShort, uriEscape } from './helpers.js'
+import * as errors from './errors.ts'
+import { getScope, isNumber, isObject, isString, makeDateLong, makeDateShort, uriEscape } from './internal/helper.ts'
+import type { ICanonicalRequest, IRequest, RequestHeaders } from './internal/type.ts'
 
 const signV4Algorithm = 'AWS4-HMAC-SHA256'
 
@@ -32,7 +32,13 @@ const signV4Algorithm = 'AWS4-HMAC-SHA256'
 //  <SignedHeaders>\n
 //  <HashedPayload>
 //
-function getCanonicalRequest(method, path, headers, signedHeaders, hashedPayload) {
+function getCanonicalRequest(
+  method: string,
+  path: string,
+  headers: RequestHeaders,
+  signedHeaders: string[],
+  hashedPayload: string,
+): ICanonicalRequest {
   if (!isString(method)) {
     throw new TypeError('method should be of type "string"')
   }
@@ -42,18 +48,19 @@ function getCanonicalRequest(method, path, headers, signedHeaders, hashedPayload
   if (!isObject(headers)) {
     throw new TypeError('headers should be of type "object"')
   }
-  if (!isArray(signedHeaders)) {
+  if (!Array.isArray(signedHeaders)) {
     throw new TypeError('signedHeaders should be of type "array"')
   }
   if (!isString(hashedPayload)) {
     throw new TypeError('hashedPayload should be of type "string"')
   }
+
   const headersArray = signedHeaders.reduce((acc, i) => {
     // Trim spaces from the value (required by V4 spec)
     const val = `${headers[i]}`.replace(/ +/g, ' ')
     acc.push(`${i.toLowerCase()}:${val}`)
     return acc
-  }, [])
+  }, [] as string[])
 
   const requestResource = path.split('?')[0]
   let requestQuery = path.split('?')[1]
@@ -65,22 +72,22 @@ function getCanonicalRequest(method, path, headers, signedHeaders, hashedPayload
     requestQuery = requestQuery
       .split('&')
       .sort()
-      .map((element) => (element.indexOf('=') === -1 ? element + '=' : element))
+      .map((element) => (!element.includes('=') ? element + '=' : element))
       .join('&')
   }
 
-  const canonical = []
-  canonical.push(method.toUpperCase())
-  canonical.push(requestResource)
-  canonical.push(requestQuery)
-  canonical.push(headersArray.join('\n') + '\n')
-  canonical.push(signedHeaders.join(';').toLowerCase())
-  canonical.push(hashedPayload)
-  return canonical.join('\n')
+  return [
+    method.toUpperCase(),
+    requestResource,
+    requestQuery,
+    headersArray.join('\n') + '\n',
+    signedHeaders.join(';').toLowerCase(),
+    hashedPayload,
+  ].join('\n')
 }
 
 // generate a credential string
-function getCredential(accessKey, region, requestDate, serviceName = 's3') {
+function getCredential(accessKey: string, region: string, requestDate?: Date, serviceName = 's3') {
   if (!isString(accessKey)) {
     throw new TypeError('accessKey should be of type "string"')
   }
@@ -94,7 +101,7 @@ function getCredential(accessKey, region, requestDate, serviceName = 's3') {
 }
 
 // Returns signed headers array - alphabetically sorted
-function getSignedHeaders(headers) {
+function getSignedHeaders(headers: RequestHeaders): string[] {
   if (!isObject(headers)) {
     throw new TypeError('request should be of type "object"')
   }
@@ -126,13 +133,13 @@ function getSignedHeaders(headers) {
   //      Is skipped for obvious reasons
 
   const ignoredHeaders = ['authorization', 'content-length', 'content-type', 'user-agent']
-  return _.map(headers, (v, header) => header)
-    .filter((header) => ignoredHeaders.indexOf(header) === -1)
+  return Object.keys(headers)
+    .filter((header) => !ignoredHeaders.includes(header))
     .sort()
 }
 
 // returns the key used for calculating signature
-function getSigningKey(date, region, secretKey, serviceName = 's3') {
+function getSigningKey(date: Date, region: string, secretKey: string, serviceName = 's3') {
   if (!isObject(date)) {
     throw new TypeError('date should be of type "object"')
   }
@@ -143,16 +150,17 @@ function getSigningKey(date, region, secretKey, serviceName = 's3') {
     throw new TypeError('secretKey should be of type "string"')
   }
   const dateLine = makeDateShort(date)
-  let hmac1 = Crypto.createHmac('sha256', 'AWS4' + secretKey)
+  const hmac1 = crypto
+      .createHmac('sha256', 'AWS4' + secretKey)
       .update(dateLine)
       .digest(),
-    hmac2 = Crypto.createHmac('sha256', hmac1).update(region).digest(),
-    hmac3 = Crypto.createHmac('sha256', hmac2).update(serviceName).digest()
-  return Crypto.createHmac('sha256', hmac3).update('aws4_request').digest()
+    hmac2 = crypto.createHmac('sha256', hmac1).update(region).digest(),
+    hmac3 = crypto.createHmac('sha256', hmac2).update(serviceName).digest()
+  return crypto.createHmac('sha256', hmac3).update('aws4_request').digest()
 }
 
 // returns the string that needs to be signed
-function getStringToSign(canonicalRequest, requestDate, region, serviceName = 's3') {
+function getStringToSign(canonicalRequest: ICanonicalRequest, requestDate: Date, region: string, serviceName = 's3') {
   if (!isString(canonicalRequest)) {
     throw new TypeError('canonicalRequest should be of type "string"')
   }
@@ -162,19 +170,15 @@ function getStringToSign(canonicalRequest, requestDate, region, serviceName = 's
   if (!isString(region)) {
     throw new TypeError('region should be of type "string"')
   }
-  const hash = Crypto.createHash('sha256').update(canonicalRequest).digest('hex')
+  const hash = crypto.createHash('sha256').update(canonicalRequest).digest('hex')
   const scope = getScope(region, requestDate, serviceName)
-  const stringToSign = []
-  stringToSign.push(signV4Algorithm)
-  stringToSign.push(makeDateLong(requestDate))
-  stringToSign.push(scope)
-  stringToSign.push(hash)
-  const signString = stringToSign.join('\n')
-  return signString
+  const stringToSign = [signV4Algorithm, makeDateLong(requestDate), scope, hash]
+
+  return stringToSign.join('\n')
 }
 
 // calculate the signature of the POST policy
-export function postPresignSignatureV4(region, date, secretKey, policyBase64) {
+export function postPresignSignatureV4(region: string, date: Date, secretKey: string, policyBase64: string): string {
   if (!isString(region)) {
     throw new TypeError('region should be of type "string"')
   }
@@ -188,11 +192,19 @@ export function postPresignSignatureV4(region, date, secretKey, policyBase64) {
     throw new TypeError('policyBase64 should be of type "string"')
   }
   const signingKey = getSigningKey(date, region, secretKey)
-  return Crypto.createHmac('sha256', signingKey).update(policyBase64).digest('hex').toLowerCase()
+  return crypto.createHmac('sha256', signingKey).update(policyBase64).digest('hex').toLowerCase()
 }
 
 // Returns the authorization header
-export function signV4(request, accessKey, secretKey, region, requestDate, serviceName = 's3') {
+export function signV4(
+  request: IRequest,
+  accessKey: string,
+  secretKey: string,
+  region: string,
+  requestDate: Date,
+  sha256sum: string,
+  serviceName = 's3',
+) {
   if (!isObject(request)) {
     throw new TypeError('request should be of type "object"')
   }
@@ -213,26 +225,41 @@ export function signV4(request, accessKey, secretKey, region, requestDate, servi
     throw new errors.SecretKeyRequiredError('secretKey is required for signing')
   }
 
-  const sha256sum = request.headers['x-amz-content-sha256']
-
   const signedHeaders = getSignedHeaders(request.headers)
   const canonicalRequest = getCanonicalRequest(request.method, request.path, request.headers, signedHeaders, sha256sum)
   const serviceIdentifier = serviceName || 's3'
   const stringToSign = getStringToSign(canonicalRequest, requestDate, region, serviceIdentifier)
   const signingKey = getSigningKey(requestDate, region, secretKey, serviceIdentifier)
   const credential = getCredential(accessKey, region, requestDate, serviceIdentifier)
-  const signature = Crypto.createHmac('sha256', signingKey).update(stringToSign).digest('hex').toLowerCase()
+  const signature = crypto.createHmac('sha256', signingKey).update(stringToSign).digest('hex').toLowerCase()
 
   return `${signV4Algorithm} Credential=${credential}, SignedHeaders=${signedHeaders
     .join(';')
     .toLowerCase()}, Signature=${signature}`
 }
 
-export function signV4ByServiceName(request, accessKey, secretKey, region, requestDate, serviceName = 's3') {
-  return signV4(request, accessKey, secretKey, region, requestDate, serviceName)
+export function signV4ByServiceName(
+  request: IRequest,
+  accessKey: string,
+  secretKey: string,
+  region: string,
+  requestDate: Date,
+  contentSha256: string,
+  serviceName = 's3',
+): string {
+  return signV4(request, accessKey, secretKey, region, requestDate, contentSha256, serviceName)
 }
+
 // returns a presigned URL string
-export function presignSignatureV4(request, accessKey, secretKey, sessionToken, region, requestDate, expires) {
+export function presignSignatureV4(
+  request: IRequest,
+  accessKey: string,
+  secretKey: string,
+  sessionToken: string | undefined,
+  region: string,
+  requestDate: Date,
+  expires: number,
+) {
   if (!isObject(request)) {
     throw new TypeError('request should be of type "object"')
   }
@@ -268,7 +295,7 @@ export function presignSignatureV4(request, accessKey, secretKey, sessionToken, 
   const credential = getCredential(accessKey, region, requestDate)
   const hashedPayload = 'UNSIGNED-PAYLOAD'
 
-  const requestQuery = []
+  const requestQuery: string[] = []
   requestQuery.push(`X-Amz-Algorithm=${signV4Algorithm}`)
   requestQuery.push(`X-Amz-Credential=${uriEscape(credential)}`)
   requestQuery.push(`X-Amz-Date=${iso8601Date}`)
@@ -292,7 +319,6 @@ export function presignSignatureV4(request, accessKey, secretKey, sessionToken, 
 
   const stringToSign = getStringToSign(canonicalRequest, requestDate, region)
   const signingKey = getSigningKey(requestDate, region, secretKey)
-  const signature = Crypto.createHmac('sha256', signingKey).update(stringToSign).digest('hex').toLowerCase()
-  const presignedUrl = request.protocol + '//' + request.headers.host + path + `&X-Amz-Signature=${signature}`
-  return presignedUrl
+  const signature = crypto.createHmac('sha256', signingKey).update(stringToSign).digest('hex').toLowerCase()
+  return request.protocol + '//' + request.headers.host + path + `&X-Amz-Signature=${signature}`
 }
